--- Vagrantfile
@@ -1,18 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-
-Vagrant.configure("2") do |config|
-
-  config.vm.box = "ubuntu/trusty64"
-
-  config.vm.provision :shell, path: "bin/vagrant.sh"
-
-  config.vm.network :forwarded_port, host: 1981, guest: 4567
-
-  config.vm.provider "virtualbox" do |v|
-    v.memory = 1024
-    v.cpus = 2
-    v.customize ["setextradata", :id, "VBoxInternal2/SharedFoldersEnableSymlinksCreate/v-root", "1"]
-  end
-
-end
