+++ package.json
@@ -3,10 +3,12 @@
   "version": "0.1.0",
   "description": "This is the seed repo for FrontEnd projects in The Greenhouse, based on React and Webpack.",
   "scripts": {
-    "build": "rimraf ./build && webpack --config ./webpack.config.prod.js -p --progress --profile --bail",
-    "ci": "export NODE_ENV=production && npm run build",
-    "develop": "export NODE_ENV=development && rimraf ./build && webpack-dev-server --config ./webpack.config.dev.js --inline --progress --port 4567",
-    "demo": "npm run build && cd ./build/ && ../node_modules/.bin/lite-server"
+    "clean": "rimraf ./build && rimraf ./reports && mkdir ./reports && mkdir ./reports/test-results",
+    "build": "webpack --config ./webpack.config.prod.js -p --progress --profile --bail",
+    "develop": "webpack-dev-server --open --config ./webpack.config.dev.js --inline --progress",
+    "serve": "webpack --config ./webpack.config.prod.js -p --progress --profile --bail && http-server ./build -o",
+    "setup": "rimraf ./src/services/credentials.ts && cp ./src/services/credentials.ts.tmpl ./src/services/credentials.ts",
+    "test": "npm run clean && export TEST_REPORT_PATH=./reports/test-results && eslint ./src/**/**/*.spec.jsx && jest --coverage --config ./jest.config.json"
   },
   "repository": {
     "type": "git",
@@ -27,18 +29,51 @@
   },
   "homepage": "https://github.com/thegreenhouseio/react-typescript-seed#readme",
   "dependencies": {
-    "react": "15.3.2",
-    "react-dom": "15.3.2"
+    "@types/react-addons-test-utils": "^0.14.18",
+    "axios": "0.15.2",
+    "bootstrap": "3.3.7",
+    "react": "15.4.0",
+    "react-addons-test-utils": "^15.5.1",
+    "react-dom": "15.4.0",
+    "react-redux": "4.4.6",
+    "react-router": "3.0.2",
+    "redux": "3.6.0"
   },
   "devDependencies": {
+    "@types/axios": "0.9.33",
+    "@types/react": "0.14.49",
+    "@types/react-dom": "0.14.18",
+    "@types/react-redux": "4.4.32",
+    "@types/react-router": "3.0.5",
+    "@types/redux": "3.6.31",
     "awesome-typescript-loader": "2.2.4",
+    "axios-mock-adapter": "1.7.1",
+    "babel-jest": "19.0.0",
+    "babel-preset-es2015": "6.22.0",
+    "babel-preset-react": "6.23.0",
+    "codelyzer": "^3.0.1",
+    "css-loader": "0.26.0",
+    "enzyme": "2.7.1",
+    "eslint": "3.18.0",
+    "eslint-plugin-react": "6.10.3",
+    "file-loader": "0.9.0",
     "html-loader": "0.4.4",
     "html-webpack-plugin": "2.22.0",
+    "http-server": "0.9.0",
+    "jest": "19.0.2",
+    "jest-junit-reporter": "1.0.1",
+    "pipeline-validate-js": "1.0.4",
+    "react-test-renderer": "^15.4.2",
+    "redux-mock-store": "^1.2.2",
     "rimraf": "2.5.4",
     "source-map-loader": "0.1.5",
+    "style-loader": "0.13.1",
+    "ts-jest": "19.0.0",
     "ts-loader": "0.8.2",
-    "typescript": "2.0.3",
+    "tslint": "^5.2.0",
+    "typescript": "2.0.10",
     "typings": "1.4.0",
+    "url-loader": "0.5.7",
     "webpack": "1.13.2",
     "webpack-dev-server": "1.16.1",
     "webpack-md5-hash": "0.0.5",
