--- src/components/hello-world/hello-world.tsx
@@ -1,14 +1,0 @@
-import * as React from "react";
-
-export interface HelloProps {
-  compiler: string;
-  framework: string;
-}
-
-export class HelloWorldComponent extends React.Component<HelloProps, {}> {
-
-  render() {
-    return <h1>Hello from {this.props.compiler} and {this.props.framework}!</h1>;
-  }
-
-}
