+++ src/vendor.ts
@@ -1,3 +1,2 @@
-//React
 import 'react';
 import 'react-dom';
