+++ src/views/home/home.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+
+export interface HelloProps {
+  compiler: string;
+  framework: string;
+}
+
+export class HomeViewComponent extends React.Component<{}, {}> {
+
+  render() {
+    return <h1>It Works!</h1>;
+  }
+
+}
