+++ tslint.json
@@ -4,12 +4,7 @@
   ],
   "rules": {
     "member-access": false,
-    "member-ordering": [
-      true,
-      "public-before-private",
-      "static-before-instance",
-      "variables-before-functions"
-    ],
+    "member-ordering": true,
     "no-any": true,
     "no-inferrable-types": false,
     "no-internal-module": true,
