--- webpack.config.dev.js
@@ -1,77 +1,0 @@
-const webpackMerge = require('webpack-merge');
-const commonConfig = require('./webpack.config.common.js');
-const DefinePlugin = require('webpack/lib/DefinePlugin');
-
-const ENV = process.env.ENV = process.env.NODE_ENV = 'development';
-const HOST = process.env.HOST || 'localhost';
-const PORT = process.env.PORT || 3000;
-const HMR = process.argv.join('').indexOf('hot') > -1;
-const METADATA = webpackMerge(commonConfig.metadata, {
-  host: HOST,
-  port: PORT,
-  ENV: ENV,
-  HMR: HMR
-});
-
-module.exports = webpackMerge(commonConfig, {
-  metadata: METADATA,
-  debug: true,
-  devtool: 'cheap-module-source-map',
-
-  output: {
-    path: __dirname + './build',
-    filename: '[name].bundle.js',
-    sourceMapFilename: '[name].map',
-    chunkFilename: '[id].chunk.js',
-    library: 'ac_[name]',
-    libraryTarget: 'var',
-  },
-
-  plugins: [
-    // NOTE: when adding more properties, make sure you include them in custom-typings.d.ts
-    new DefinePlugin({
-      'ENV': JSON.stringify(METADATA.ENV),
-      'HMR': METADATA.HMR,
-      'process.env': {
-        'ENV': JSON.stringify(METADATA.ENV),
-        'NODE_ENV': JSON.stringify(METADATA.ENV),
-        'HMR': METADATA.HMR,
-      }
-    }),
-  ],
-
-  tslint: {
-    emitErrors: false,
-    failOnHint: false,
-    resourcePath: 'src'
-  },
-
-  devServer: {
-    port: METADATA.port,
-    host: METADATA.host,
-    historyApiFallback: true,
-    watchOptions: {
-      aggregateTimeout: 300,
-      poll: 1000
-    },
-    outputPath: __dirname + '/build',
-    proxy:{
-      '/api/*': {
-        target: 'http://<your-domain>',
-        secure: false,
-        changeOrigin: true
-      }
-    }
-  },
-
-  //TODO is this needed?
-  node: {
-    global: 'window',
-    crypto: 'empty',
-    process: true,
-    module: false,
-    clearImmediate: false,
-    setImmediate: false
-  }
-
-});
